(PC-22628)[BSR] refactor: replace info phone icon and refactor radioButtonWithImage with SvgIcon component
