+++ pro/src/screens/OfferType/OfferType.tsx
@@ -24,16 +24,16 @@ import {
 import { getOfferIndividualUrl } from 'core/Offers/utils/getOfferIndividualUrl'
 import useAnalytics from 'hooks/useAnalytics'
 import useNotification from 'hooks/useNotification'
-import { ReactComponent as CalendarCheckIcon } from 'icons/ico-calendar-check.svg'
-import { ReactComponent as CaseIcon } from 'icons/ico-case.svg'
-import { ReactComponent as DateIcon } from 'icons/ico-date.svg'
-import { ReactComponent as DuplicateOfferIcon } from 'icons/ico-duplicate-offer.svg'
-import { ReactComponent as NewOfferIcon } from 'icons/ico-new-offer.svg'
-import { ReactComponent as TemplateOfferIcon } from 'icons/ico-template-offer.svg'
-import { ReactComponent as ThingIcon } from 'icons/ico-thing.svg'
-import { ReactComponent as VirtualEventIcon } from 'icons/ico-virtual-event.svg'
-import { ReactComponent as VirtualThingIcon } from 'icons/ico-virtual-thing.svg'
-import { ReactComponent as PhoneIcon } from 'icons/info-phone.svg'
+import CalendarCheckIcon from 'icons/ico-calendar-check.svg'
+import CaseIcon from 'icons/ico-case.svg'
+import DateIcon from 'icons/ico-date.svg'
+import DuplicateOfferIcon from 'icons/ico-duplicate-offer.svg'
+import NewOfferIcon from 'icons/ico-new-offer.svg'
+import TemplateOfferIcon from 'icons/ico-template-offer.svg'
+import VirtualEventIcon from 'icons/ico-virtual-event.svg'
+import VirtualThingIcon from 'icons/ico-virtual-thing.svg'
+import PhoneIcon from 'icons/stroke-info-phone.svg'
+import ThingIcon from 'icons/stroke-thing.svg'
 import { getFilteredCollectiveOffersAdapter } from 'pages/CollectiveOffers/adapters'
 import { Banner } from 'ui-kit'
 import RadioButtonWithImage from 'ui-kit/RadioButtonWithImage'
@@ -190,7 +190,7 @@ const OfferType = (): JSX.Element => {
               <FormLayout.Row inline>
                 <RadioButtonWithImage
                   name="offerType"
-                  Icon={PhoneIcon}
+                  icon={PhoneIcon}
                   isChecked={values.offerType === OFFER_TYPES.INDIVIDUAL_OR_DUO}
                   label="Au grand public"
                   onChange={handleChange}
@@ -199,7 +199,7 @@ const OfferType = (): JSX.Element => {
                 />
                 <RadioButtonWithImage
                   name="offerType"
-                  Icon={CaseIcon}
+                  icon={CaseIcon}
                   isChecked={values.offerType === OFFER_TYPES.EDUCATIONAL}
                   label="À un groupe scolaire"
                   onChange={handleChange}
@@ -220,7 +220,7 @@ const OfferType = (): JSX.Element => {
                   <FormLayout.Row inline mdSpaceAfter>
                     <RadioButtonWithImage
                       name="collectiveOfferSubtype"
-                      Icon={CalendarCheckIcon}
+                      icon={CalendarCheckIcon}
                       isChecked={
                         values.collectiveOfferSubtype ===
                         COLLECTIVE_OFFER_SUBTYPE.COLLECTIVE
@@ -235,7 +235,7 @@ const OfferType = (): JSX.Element => {
                   <FormLayout.Row inline mdSpaceAfter>
                     <RadioButtonWithImage
                       name="collectiveOfferSubtype"
-                      Icon={TemplateOfferIcon}
+                      icon={TemplateOfferIcon}
                       isChecked={
                         values.collectiveOfferSubtype ===
                         COLLECTIVE_OFFER_SUBTYPE.TEMPLATE
@@ -257,7 +257,7 @@ const OfferType = (): JSX.Element => {
                   <RadioButtonWithImage
                     className={styles['individual-radio-button']}
                     name="individualOfferSubtype"
-                    Icon={ThingIcon}
+                    icon={ThingIcon}
                     isChecked={
                       values.individualOfferSubtype ===
                       INDIVIDUAL_OFFER_SUBTYPE.PHYSICAL_GOOD
@@ -274,7 +274,7 @@ const OfferType = (): JSX.Element => {
                   <RadioButtonWithImage
                     className={styles['individual-radio-button']}
                     name="individualOfferSubtype"
-                    Icon={VirtualThingIcon}
+                    icon={VirtualThingIcon}
                     isChecked={
                       values.individualOfferSubtype ===
                       INDIVIDUAL_OFFER_SUBTYPE.VIRTUAL_GOOD
@@ -291,7 +291,7 @@ const OfferType = (): JSX.Element => {
                   <RadioButtonWithImage
                     className={styles['individual-radio-button']}
                     name="individualOfferSubtype"
-                    Icon={DateIcon}
+                    icon={DateIcon}
                     isChecked={
                       values.individualOfferSubtype ===
                       INDIVIDUAL_OFFER_SUBTYPE.PHYSICAL_EVENT
@@ -308,7 +308,7 @@ const OfferType = (): JSX.Element => {
                   <RadioButtonWithImage
                     className={styles['individual-radio-button']}
                     name="individualOfferSubtype"
-                    Icon={VirtualEventIcon}
+                    icon={VirtualEventIcon}
                     isChecked={
                       values.individualOfferSubtype ===
                       INDIVIDUAL_OFFER_SUBTYPE.VIRTUAL_EVENT
@@ -334,7 +334,7 @@ const OfferType = (): JSX.Element => {
                   <FormLayout.Row inline mdSpaceAfter>
                     <RadioButtonWithImage
                       name="collectiveOfferSubtypeDuplicate"
-                      Icon={NewOfferIcon}
+                      icon={NewOfferIcon}
                       isChecked={
                         values.collectiveOfferSubtypeDuplicate ===
                         COLLECTIVE_OFFER_SUBTYPE_DUPLICATE.NEW_OFFER
@@ -349,7 +349,7 @@ const OfferType = (): JSX.Element => {
                     <RadioButtonWithImage
                       name="collectiveOfferSubtypeDuplicate"
                       transparent
-                      Icon={DuplicateOfferIcon}
+                      icon={DuplicateOfferIcon}
                       isChecked={
                         values.collectiveOfferSubtypeDuplicate ===
                         COLLECTIVE_OFFER_SUBTYPE_DUPLICATE.DUPLICATE
